initial commit
contains the base code for storing local secrets on a system. it is
fairly bare bones but will probably add more to it later.
added in some tests to make sure that the (de)serialization processes
work as they are supposed to.
encrypting the manager file is on the list of things to do but right now
this will be fine.